LayoutTests/imported/w3c/web-platform-tests/content-security-policy/child-src/child-src-worker-blocked.sub-expected.txt

This test used to check the child-src csp controlling worker creation. This behaviour has been deprecated but it's still supported until the transition is done. This still tests that behaviour but we need to go through extra hoops to make sure 'script-src' does not affect the result in any way (for instance by allowing 'self').


FAIL Should throw a securitypolicyviolation event assert_equals: expected "worker-src" but got "child-src 'none'"
PASS Should block worker because it does not match any directive including the deprecated 'child-src'
